+++ src/extension.ts
@@ -6,26 +6,32 @@ interface StringObject {
 }
 
 // Constants
+const BLOCK_COMMENT_DELIMITERS: StringObject = {"/*": "*/"};
+const DECORATION_TYPE = vscode.window.createTextEditorDecorationType({});
 const EM_PER_CHAR = 0.55;
 const ENABLED = false;
+const ESCAPABLE_CHARS = '"\'`';
+const INLINE_COMMENT_DELIMTERS: string[] = ["//"];
 const MAX_COMMENT_LINE_LENGTH = 80;
-const MULTI_LINE_COMMENT_STRINGS: StringObject = { "/*": "*/" };
-const SINGLE_LINE_COMMENT_STRINGS = ["//"];
+const MULTI_LINE_STRING_DELIMITERS: string[] = [];
+const SINGLE_LINE_STRING_DELIMITERS: string[] = ['"', "'", "`"];
 const STYLE = "1px solid green";
-const DECORATION_TYPE: vscode.TextEditorDecorationType = vscode.window.createTextEditorDecorationType({});
 
 // Global settings
+let blockCommentDelimiters = BLOCK_COMMENT_DELIMITERS;
 let emPerChar = EM_PER_CHAR;
 let enabled = ENABLED;
+let escapableChars = ESCAPABLE_CHARS;
+let inlineCommentDelimiters = INLINE_COMMENT_DELIMTERS;
 let maxCommentLineLength = MAX_COMMENT_LINE_LENGTH;
-let multiLineCommentStrings = MULTI_LINE_COMMENT_STRINGS;
-let singleLineCommentStrings = SINGLE_LINE_COMMENT_STRINGS;
+let multiLineStringDelimiters = MULTI_LINE_STRING_DELIMITERS;
+let singleLineStringDelimiters = SINGLE_LINE_STRING_DELIMITERS;
 let style = STYLE;
 
 // Comment status
-let inSingleLineComment = -1;
-let inMultiLineComment: string | null = null;
-let multiLineCommentStart = 0;
+let blockCommentStart = 0;
+let inBlockComment: string | null = null;
+let inMultiLineString = -1;
 
 function activate(_: vscode.ExtensionContext) {
 	drawCommentLines();
@@ -37,69 +43,84 @@ function activate(_: vscode.ExtensionContext) {
 
 function updateSettings(languageId: string) {
 	for (let getConfig of [vscode.workspace.getConfiguration('comment-rulers').get, vscode.workspace.getConfiguration('comment-rulers', { languageId: languageId }).get]) {
+		blockCommentDelimiters = getConfig<StringObject>('blockCommentDelimiters', BLOCK_COMMENT_DELIMITERS);
 		emPerChar = getConfig<number>('emPerChar', EM_PER_CHAR);
 		enabled = getConfig<boolean>('enabled', ENABLED);
+		escapableChars = getConfig<string>('escapableChars', ESCAPABLE_CHARS);
+		inlineCommentDelimiters = getConfig<string[]>('inlineCommentDelimiters', INLINE_COMMENT_DELIMTERS);
 		maxCommentLineLength = getConfig<number>('maxCommentLineLength', MAX_COMMENT_LINE_LENGTH);
-		multiLineCommentStrings = getConfig<StringObject>('multiLineCommentStrings', MULTI_LINE_COMMENT_STRINGS);
-		singleLineCommentStrings = getConfig<string[]>('singleLineCommentStrings', SINGLE_LINE_COMMENT_STRINGS);
+		multiLineStringDelimiters = getConfig<string[]>('multiLineStringDelimiters', MULTI_LINE_STRING_DELIMITERS);
+		singleLineStringDelimiters = getConfig<string[]>('singleLineStringDelimiters', SINGLE_LINE_STRING_DELIMITERS);
 		style = getConfig<string>('style', STYLE);
 	}
 }
 
 function parseLine(line: string) {
-	let start = line.length;
-	let multiLineCommentEndMatch = "";
-	let singleLineCommentMatches = new Array(singleLineCommentStrings.length).fill("");
-	let multiLineCommentStartMatches: StringObject = {};
-	for (const key in multiLineCommentStrings) {
-		multiLineCommentStartMatches[key] = "";
+	let commentStart = Infinity;
+	let inSingleLineString = -1;
+	if (inBlockComment !== null) {
+		commentStart = blockCommentStart;
 	}
 	for (let i = 0; i < line.length; i++) {
-		const char = line[i];
-		let newStart;
-		for (let j = 0; j < singleLineCommentMatches.length; j++) {
-			singleLineCommentMatches[j] += char;
-			newStart = i - singleLineCommentMatches[j].length + 1;
-			if (singleLineCommentMatches[j] === singleLineCommentStrings[j]) {
-				if (newStart < start) {
-					inSingleLineComment = j;
-					start = newStart;
-				} else if (inMultiLineComment === null) {
-					inSingleLineComment = j;
-				}
+		if (line[i] === "\\" && escapableChars.includes(line[++i])) {
+			continue;
+		} // else
+		if (inBlockComment !== null) {
+			if (line.substring(i, i + blockCommentDelimiters[inBlockComment].length) === blockCommentDelimiters[inBlockComment]) {
+				inBlockComment = null;
 			}
-			if (!singleLineCommentStrings[j].startsWith(singleLineCommentMatches[j])) {
-				singleLineCommentMatches[j] = "";
+			continue;
+		} // else
+		if (inMultiLineString !== -1) {
+			if (line.substring(i, i + multiLineStringDelimiters[inMultiLineString].length) === multiLineStringDelimiters[inMultiLineString]) {
+				inMultiLineString = -1;
 			}
-		}
-		for (const key in multiLineCommentStartMatches) {
-			multiLineCommentStartMatches[key] += char;
-			newStart = i - multiLineCommentStartMatches[key].length + 1;
-			if (multiLineCommentStartMatches[key] === key) {
-				if (newStart < start) {
-					inMultiLineComment = key;
-					multiLineCommentStart = start = newStart;
-				} else if (inSingleLineComment === -1) {
-					inMultiLineComment = key;
-					multiLineCommentStart = newStart;
-				}
+			continue;
+		} // else
+		if (inSingleLineString !== -1) {
+			if (line.substring(i, i + singleLineStringDelimiters[inSingleLineString].length) === singleLineStringDelimiters[inSingleLineString]) {
+				inSingleLineString = -1;
 			}
-			if (!key.startsWith(multiLineCommentStartMatches[key])) {
-				multiLineCommentStartMatches[key] = "";
+			continue;
+		} // else
+		for (const value of inlineCommentDelimiters) {
+			if (line.substring(i, i + value.length) === value) {
+				return Math.min(i, commentStart);
 			}
 		}
-		multiLineCommentEndMatch += char;
-		if (inMultiLineComment !== null && multiLineCommentEndMatch === multiLineCommentStrings[inMultiLineComment]) {
-			if (multiLineCommentStart < start) {
-				start = multiLineCommentStart;
-			}
-			inMultiLineComment = null;
+		for (const key in blockCommentDelimiters) {
+			if (line.substring(i, i + key.length) !== key) {
+				continue;
+			} // else
+			inBlockComment = key;
+			blockCommentStart = i;
+			commentStart = Math.min(i, commentStart);
+			break;
 		}
-		if (inMultiLineComment !== null && !multiLineCommentStrings[inMultiLineComment].startsWith(multiLineCommentEndMatch)) {
-			multiLineCommentEndMatch = "";
+		if (inBlockComment !== null) {
+			continue;
+		} // else
+		for (let j = 0; j < multiLineStringDelimiters.length; j++) {
+			const value = multiLineStringDelimiters[j];
+			if (line.substring(i, i + value.length) !== value) {
+				continue;
+			}
+			inMultiLineString = j;
+			break;
+		}
+		if (inMultiLineString !== -1) {
+			continue;
+		} // else
+		for (let j = 0; j < singleLineStringDelimiters.length; j++) {
+			const value = singleLineStringDelimiters[j];
+			if (line.substring(i, i + value.length) !== value) {
+				continue;
+			}
+			inSingleLineString = j;
+			break;
 		}
 	}
-	return start;
+	return commentStart;
 }
 
 function drawCommentLines() {
@@ -111,18 +132,15 @@ function drawCommentLines() {
 		vscode.window.activeTextEditor.setDecorations(DECORATION_TYPE, []);
 		return;
 	}
-	const document = vscode.window.activeTextEditor.document;
+	blockCommentStart = 0;
 	let decorations: vscode.DecorationOptions[] = [];
-	inMultiLineComment = null;
-	multiLineCommentStart = 0;
+	const document = vscode.window.activeTextEditor.document;
+	inBlockComment = null;
+	inMultiLineString = -1;
 	for (let i = 0; i < document.lineCount; i++) {
 		const line = document.lineAt(i).text;
-		const end = line.length;
-		inSingleLineComment = -1;
-		let start = parseLine(line);
-		if (inMultiLineComment !== null && multiLineCommentStart < start) {
-			start = multiLineCommentStart;
-		} else if (start === end) {
+		let start = Math.max(0, parseLine("\n" + line + "\n") - 1);
+		if (start === Infinity) {
 			continue;
 		}
 		const decoration = {
@@ -131,7 +149,7 @@ function drawCommentLines() {
 				after: {
 					contentText: "",
 					border: style,
-					margin: `0 0 0 ${Math.max(0, emPerChar * (start + maxCommentLineLength - end))}em`
+					margin: `0 0 0 ${emPerChar * Math.max(0, start + maxCommentLineLength - line.length)}em`
 				}
 			},
 			rangeBehavior: vscode.DecorationRangeBehavior.ClosedClosed