+++ CHANGELOG.md
@@ -9,3 +9,9 @@ All notable changes to the "comment-rulers" extension will be documented in this
 ## 1.0.1 (Apr 14. 2023)
 
 Fixed Multi-line comments are messed up
+
+### 1.0.2 (Apr 16. 2023)
+
+Detecting strings (to ingore included comment delimiters).
+Escaping delimiters
+Fixed features.gif